refactor: changed typo from infraestructure to infrastructure
