+++ Infrastructure/IActivityInfraestructure.cs
@@ -1,7 +1,7 @@
-﻿using Infraestructure.DataClass;
+﻿using Infrastructure.DataClass;
 using Infrastructure.Model;
 
-namespace Infraestructure;
+namespace Infrastructure;
 
 public interface IActivityInfraestructure
 {
